ci {gmodels}	R Documentation

Compute Confidence Intervals

Description

Compute and display confidence intervals for model estimates. Methods are provided for the mean of a numeric vector ci.default, the probability of a binomial vector ci.binom, and for lm, lme, and mer objects are provided.

Usage

ci(x, confidence = 0.95, alpha = 1 - confidence, ...)

## S3 method for class 'numeric'
ci(x, confidence = 0.95, alpha = 1 - confidence, na.rm = FALSE, ...)

Arguments

x
object from which to compute confidence intervals.

confidence
confidence level. Defaults to 0.95.

alpha
type one error rate. Defaults to 1.0-confidence

...
Arguments for methods

na.rm
logical indicating whether missing values should be removed.

Value

vector or matrix with one row per model parameter and elements/columns Estimate, CI lower, CI upper, Std. Error, DF (for lme objects only), and p-value.

Author(s)

Gregory R. Warnes greg@warnes.net

See Also

stats::confint(), stats::lm(), stats::summary.lm()

Examples

# mean and confidence interval
ci( rnorm(10) )

# binomial proportion and exact confidence interval
b <- rbinom( prob=0.75, size=1, n=20 )
ci.binom(b) # direct call
class(b) <- 'binom'
ci(b)       # indirect call

# confidence intervals for regression parameteres
data(state)
reg  <-  lm(Area ~ Population, data=as.data.frame(state.x77))
ci(reg)

[Package gmodels version 2.19.1 Index]